+++ sdk/domainregistration/Azure.Provisioning.DomainRegistration/tests/BasicDomainRegistrationTests.cs
@@ -60,6 +60,7 @@ public class BasicDomainRegistrationTests
     }
 
     [Test]
+    [Ignore("Restore the test when next version of azure.provisioning that properly supports datetime format releases.")]
     public async Task CreateAppServiceDomain()
     {
         await using Trycep test = CreateAppServiceDomainTest();
@@ -131,7 +132,7 @@ public class BasicDomainRegistrationTests
                     'agreement-key'
                   ]
                   agreedBy: '192.0.2.1'
-                  agreedAt: '2026-01-01T00:00:00.0000000+00:00'
+                  agreedAt: '2026-01-01T00:00:00.0000000Z'
                 }
               }
             }