+++ sdk/provisioning/Azure.Provisioning/src/Primitives/ProvisionableConstruct.cs
@@ -346,12 +346,21 @@ public abstract class ProvisionableConstruct : Provisionable, IBicepValue
         string[]? bicepPath,
         bool isOutput = false,
         bool isRequired = false)
+        => DefineListProperty<T>(propertyName, bicepPath, isOutput, isRequired, format: null);
+
+    protected BicepList<T> DefineListProperty<T>(
+        string propertyName,
+        string[]? bicepPath,
+        bool isOutput,
+        bool isRequired,
+        string? format)
     {
         BicepList<T> values =
             new(new BicepValueReference(this, propertyName, bicepPath), values: null) // we call this ctor to initialize an "uninitialized" list
             {
                 _isOutput = isOutput,
-                _isRequired = isRequired
+                _isRequired = isRequired,
+                Format = format
             };
         ProvisionableProperties[propertyName] = values;
         return values;
@@ -371,12 +380,21 @@ public abstract class ProvisionableConstruct : Provisionable, IBicepValue
         string[]? bicepPath,
         bool isOutput = false,
         bool isRequired = false)
+        => DefineDictionaryProperty<T>(propertyName, bicepPath, isOutput, isRequired, format: null);
+
+    protected BicepDictionary<T> DefineDictionaryProperty<T>(
+        string propertyName,
+        string[]? bicepPath,
+        bool isOutput,
+        bool isRequired,
+        string? format)
     {
         BicepDictionary<T> values =
             new(new BicepValueReference(this, propertyName, bicepPath), values: null) // we call this ctor to initialize an "uninitialized" dictionary
             {
                 _isOutput = isOutput,
-                _isRequired = isRequired
+                _isRequired = isRequired,
+                Format = format
             };
         ProvisionableProperties[propertyName] = values;
         return values;